minor fixes;

+++ src/driver/oracle/OracleQueryRunner.ts
@@ -514,7 +514,7 @@ export class OracleQueryRunner extends BaseQueryRunner implements QueryRunner {
         if (!oldColumn)
             throw new Error(`Column "${oldTableColumnOrName}" was not found in the "${table.name}" table.`);
 
-        if (newColumn.isGenerated !== oldColumn.isGenerated) {
+        if (newColumn.isGenerated !== oldColumn.isGenerated && newColumn.generationStrategy === "increment") {
             throw new Error(`Changing column's "isGenerated" property is not supported in Oracle driver. Drop column and recreate it with a new "isGenerated" property instead.`);
         }
